add real data test
