+++ pyproject.toml
@@ -54,6 +54,7 @@ dependencies = [
 
 [dependency-groups]
 dev = [
+    "imagecodecs>=2025.3.30",
     "pytest>=8.0",
     "pytest-cov>=5.0",
     "ruff>=0.4",
